pyproject.toml
[project]
name = "kunshort-django-payment"
version = "0.1.2"
description = "Reusable Django payment app with a unified service interface across multiple payment providers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "requests>=2.32",
    "celery>=5.3",
    "django-redis>=5.4",
    "pillow>=10.0",
    "drf-spectacular>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-django>=4.8",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src"]
python_files = ["test_*.py", "*_test.py", "tests.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kunshort_payment"]
